src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadow-tree-styles-select-host-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
Test for issue 247275: shadow tree styles selects shadow host.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test that tag selectors in a shadow tree cannot match without :host.
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
